CO-1703 Refactor
