+++ src/config.ts
@@ -1,7 +1,16 @@
-export type DeploymentType = "npm" | "make";
+const input = (name: string) => {
+  const envName = `INPUT_${name}`.toUpperCase().replace(" ", "_");
+  const value = process.env[envName];
+  if (typeof value === "undefined") {
+    throw new Error(`Input ${name} was not provided`);
+  }
+  return value;
+};
 
 export const config = {
   statusCheckContext: "QA",
-  preProductionEnvironment: "integration",
-  deploymentType: "npm" as DeploymentType, // npm | make
+  productionEnvironment: input("production-environment"),
+  preProductionEnvironment: input("pre-production-environment"),
+  deployCommand: input("deploy"),
+  releaseCommand: input("release"),
 };
